requirements.txt
sanic==22.12.0
python-dotenv==1.0.0
diffusers==0.14.0
torch==2.0.0
scipy==1.10.1
transformers==4.27.3
accelerate==0.18.0